+++ src/main/java/com/iridium/iridiumskyblock/bank/CrystalsBankItem.java
@@ -6,6 +6,7 @@ import com.iridium.iridiumskyblock.IridiumSkyblock;
 import com.iridium.iridiumskyblock.database.Island;
 import com.iridium.iridiumskyblock.database.IslandBank;
 import com.iridium.iridiumskyblock.database.User;
+import com.iridium.iridiumskyblock.gui.IslandBankGUI;
 import lombok.NoArgsConstructor;
 import org.bukkit.entity.Player;
 
@@ -22,7 +23,7 @@ public class CrystalsBankItem extends BankItem {
      * The default constructor.
      *
      * @param defaultAmount The default withdrawal amount of this item
-     * @param item          The Item which represents this bank item in the {@link com.iridium.iridiumskyblock.gui.BankGUI}
+     * @param item          The Item which represents this bank item in the {@link IslandBankGUI}
      */
     public CrystalsBankItem(double defaultAmount, Item item) {
         super("crystals", "Crystals", defaultAmount, true, item);
