+++ src/main/java/com/iridium/iridiumskyblock/commands/LogsCommand.java
@@ -4,13 +4,14 @@ import com.iridium.iridiumcore.utils.StringUtils;
 import com.iridium.iridiumskyblock.IridiumSkyblock;
 import com.iridium.iridiumskyblock.database.Island;
 import com.iridium.iridiumskyblock.database.User;
-import com.iridium.iridiumskyblock.gui.LogsGUI;
+import com.iridium.iridiumskyblock.gui.IslandLogsGUI;
+import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
+
 import java.time.Duration;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
-import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
 
 /**
  * Command which opens the Island bank GUI.
@@ -41,7 +42,7 @@ public class LogsCommand extends Command {
             return false;
         }
 
-        player.openInventory(new LogsGUI(island.get()).getInventory());
+        player.openInventory(new IslandLogsGUI(island.get()).getInventory());
         return true;
     }
