+++ src/main/java/com/iridium/iridiumskyblock/gui/BlockValueGUI.java
@@ -32,7 +32,7 @@ public class BlockValueGUI extends GUI {
      * @param type The type of valuable block shown in this GUI
      */
     public BlockValueGUI(BlockValueType type) {
-        super(IridiumSkyblock.getInstance().getInventories().blockValue, null);
+        super(IridiumSkyblock.getInstance().getInventories().blockValue);
         this.guiType = type;
     }
