+++ src/main/java/com/iridium/iridiumskyblock/gui/IslandMembersGUI.java
@@ -2,11 +2,9 @@ package com.iridium.iridiumskyblock.gui;
 
 import com.iridium.iridiumcore.utils.InventoryUtils;
 import com.iridium.iridiumcore.utils.ItemStackUtils;
-import com.iridium.iridiumcore.utils.Placeholder;
 import com.iridium.iridiumskyblock.IridiumSkyblock;
 import com.iridium.iridiumskyblock.PlaceholderBuilder;
 import com.iridium.iridiumskyblock.database.Island;
-import com.iridium.iridiumskyblock.database.IslandTrusted;
 import com.iridium.iridiumskyblock.database.User;
 import org.bukkit.Bukkit;
 import org.bukkit.event.inventory.InventoryClickEvent;
@@ -14,14 +12,13 @@ import org.bukkit.inventory.Inventory;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
 
 /**
- * GUI which displays all trusted users of an Island.
+ * GUI which displays all members of an Island and allows quick rank management.
  */
-public class TrustedGUI extends GUI {
+public class IslandMembersGUI extends IslandGUI {
 
     private final Map<Integer, User> members;
 
@@ -30,24 +27,22 @@ public class TrustedGUI extends GUI {
      *
      * @param island The Island this GUI belongs to
      */
-    public TrustedGUI(@NotNull Island island) {
-        super(IridiumSkyblock.getInstance().getInventories().trustedGUI, island);
+    public IslandMembersGUI(@NotNull Island island) {
+        super(IridiumSkyblock.getInstance().getInventories().membersGUI, island);
         this.members = new HashMap<>();
     }
 
     @Override
     public void addContent(Inventory inventory) {
         inventory.clear();
-        InventoryUtils.fillInventory(inventory, IridiumSkyblock.getInstance().getInventories().trustedGUI.background);
-
-        List<IslandTrusted> islandTrustedList = IridiumSkyblock.getInstance().getDatabaseManager().getIslandTrustedTableManager().getEntries(getIsland());
+        members.clear();
+        InventoryUtils.fillInventory(inventory, IridiumSkyblock.getInstance().getInventories().membersGUI.background);
         AtomicInteger slot = new AtomicInteger(0);
-        for (IslandTrusted islandTrusted : islandTrustedList) {
+
+        for (User user : getIsland().getMembers()) {
             int itemSlot = slot.getAndIncrement();
-            List<Placeholder> placeholderList = new PlaceholderBuilder().applyPlayerPlaceholders(islandTrusted.getUser()).applyIslandPlaceholders(getIsland()).build();
-            placeholderList.add(new Placeholder("trustee", islandTrusted.getTruster().getName()));
-            inventory.setItem(itemSlot, ItemStackUtils.makeItem(IridiumSkyblock.getInstance().getInventories().trustedGUI.item, placeholderList));
-            members.put(itemSlot, islandTrusted.getUser());
+            inventory.setItem(itemSlot, ItemStackUtils.makeItem(IridiumSkyblock.getInstance().getInventories().membersGUI.item, new PlaceholderBuilder().applyPlayerPlaceholders(user).applyIslandPlaceholders(getIsland()).build()));
+            members.put(itemSlot, user);
         }
     }
 
@@ -61,8 +56,23 @@ public class TrustedGUI extends GUI {
     public void onInventoryClick(InventoryClickEvent event) {
         if (members.containsKey(event.getSlot())) {
             User user = members.get(event.getSlot());
-            String command = IridiumSkyblock.getInstance().getCommands().unTrustCommand.aliases.get(0);
-            Bukkit.getServer().dispatchCommand(event.getWhoClicked(), "is " + command + " " + user.getName());
+
+            String command;
+            switch (event.getClick()) {
+                case LEFT:
+                    command = IridiumSkyblock.getInstance().getCommands().demoteCommand.aliases.get(0);
+                    break;
+                case RIGHT:
+                    command = IridiumSkyblock.getInstance().getCommands().promoteCommand.aliases.get(0);
+                    break;
+                default:
+                    return;
+            }
+
+            if (command != null) {
+                Bukkit.getServer().dispatchCommand(event.getWhoClicked(), "is " + command + " " + user.getName());
+            }
+
             addContent(event.getInventory());
         }
     }
