+++ src/main/java/com/iridium/iridiumskyblock/gui/IslandTopGUI.java
@@ -27,7 +27,7 @@ public class IslandTopGUI extends GUI {
      * The default constructor.
      */
     public IslandTopGUI() {
-        super(IridiumSkyblock.getInstance().getInventories().islandTopGUI, null);
+        super(IridiumSkyblock.getInstance().getInventories().islandTopGUI);
     }
 
     @Override
