+++ src/main/java/com/iridium/iridiumskyblock/gui/SchematicGUI.java
@@ -19,7 +19,7 @@ public abstract class SchematicGUI extends GUI {
     private final Map<Integer, Schematics.SchematicConfig> schematics = new HashMap<>();
 
     public SchematicGUI() {
-        super(IridiumSkyblock.getInstance().getInventories().islandSchematicGUI, null);
+        super(IridiumSkyblock.getInstance().getInventories().islandSchematicGUI);
     }
 
     @Override
